creating shallow and deep objects in java

+++ src/Chapter10/Employee.java
@@ -33,5 +33,5 @@ public abstract class Employee {
         );
     }
 
-    public abstract double earnings() ;
+    public abstract int earnings() ;
 }
